requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
selenium
webdriver-manager
beautifulsoup4
requests
python-multipart
python-dotenv
paddlepaddle
paddleocr
pyzbar
opencv-python
Pillow
passlib[bcrypt]
python-jose[cryptography]
slowapi
google-generativeai
yt-dlp
openai-whisper
